requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
langchain==0.1.9
langchain-google-genai==0.0.11
google-generativeai==0.3.2
python-dotenv==1.0.1
pydantic==2.6.1